+++ packages/eslint-plugin/src/rules/no-floating-promises.ts
@@ -24,14 +24,15 @@ export default util.createRule<Options, MessageId>({
       description: 'Requires Promise-like values to be handled appropriately',
       category: 'Best Practices',
       recommended: false,
+      suggestion: true,
       requiresTypeChecking: true,
     },
     messages: {
-      floating: 'Promises must be handled appropriately',
+      floating: 'Promises must be handled appropriately.',
       floatingVoid:
         'Promises must be handled appropriately' +
-        ' or explicitly marked as ignored with the `void` operator',
-      floatingFixVoid: 'Add void operator to ignore',
+        ' or explicitly marked as ignored with the `void` operator.',
+      floatingFixVoid: 'Add void operator to ignore.',
     },
     schema: [
       {