+++ packages/eslint-plugin/src/rules/no-unnecessary-qualifier.ts
@@ -179,7 +179,7 @@ export default util.createRule({
       TSQualifiedName(node: TSESTree.TSQualifiedName): void {
         visitNamespaceAccess(node, node.left, node.right);
       },
-      'MemberExpression[computed=false]': function(
+      'MemberExpression[computed=false]': function (
         node: TSESTree.MemberExpression,
       ): void {
         const property = node.property as TSESTree.Identifier;