+++ packages/eslint-plugin/src/rules/switch-exhaustiveness-check.ts
@@ -17,13 +17,14 @@ export default createRule({
       description: 'Exhaustiveness checking in switch with union type',
       category: 'Best Practices',
       recommended: false,
+      suggestion: true,
       requiresTypeChecking: true,
     },
     schema: [],
     messages: {
       switchIsNotExhaustive:
         'Switch is not exhaustive. Cases not matched: {{missingBranches}}',
-      addMissingCases: 'Add branches for missing cases',
+      addMissingCases: 'Add branches for missing cases.',
     },
   },
   defaultOptions: [],
@@ -129,7 +130,7 @@ export default createRule({
             missingBranches: missingBranchTypes
               .map(missingType =>
                 isTypeFlagSet(missingType, ts.TypeFlags.ESSymbolLike)
-                  ? `typeof ${missingType.symbol.escapedName}`
+                  ? `typeof ${missingType.getSymbol()?.escapedName}`
                   : checker.typeToString(missingType),
               )
               .join(' | '),