+++ packages/typescript-estree/typings/typescript.d.ts
@@ -1,0 +1,8 @@
+import 'typescript';
+
+declare module 'typescript' {
+  interface SourceFile {
+    // this is marked as internal to typescript
+    externalModuleIndicator?: Node;
+  }
+}